+++ tempolivre/calculadoraprojeto.py
@@ -1,6 +1,6 @@
 def cauculo():
-    numero1=int(input("numero1:"))
-    numero2=int(input("numero2:"))
+    numero1=float(input("numero1:"))
+    numero2=float(input("numero2:"))
     op=input("operação:")
     rei=input("reiniciar?")
     soma= numero1 + numero2
