+++ tempolivre/nada.py
@@ -1,5 +1,5 @@
 #quando usar o input se não for para caracteres é nescessario usar, int(), em caso de numeros inteiros e ,float(), em caso de decimais
-numero= int(input("Quanto:"))
+numero= float(input("Quanto:"))
 nome=input("Oque:")
 if numero==10 and nome=="batatas":
     print("Você tem 10 batatas")
